Fix potential truncation in FUnversionedPropertySerializer constructor
#rb Johan.Torp
#jira none
#rnx
#preflight 643935f98e01968448188986

- It is expected that the value of GetIntNum will fit in a uint8 so we can just cast.

[CL 25194695 by paul chipchase in ue5-main branch]

+++ Engine/Source/Runtime/CoreUObject/Private/Serialization/UnversionedPropertySerialization.cpp
@@ -50,7 +50,7 @@ public:
 		, Offset(Property->GetOffset_ForInternal() + Property->ElementSize * InArrayIndex)
 		, bSerializeAsInteger(CanSerializeAsInteger(Property))
 		, IntType(GetIntType(Property->GetMinAlignment()))
-		, FastZeroIntNum(CanSerializeAsZero(InProperty, IntType) ? GetIntNum(InProperty, IntType) : uint8(0))
+		, FastZeroIntNum(CanSerializeAsZero(InProperty, IntType) ? static_cast<uint8>(GetIntNum(InProperty, IntType)) : uint8(0))
 #else
 		, ArrayIndex(InArrayIndex)
 #endif
